[quakeio] Use size_t for Qread and Qwrite
This makes them more compatible with fread and fwrite, and probably more
importantly, support larger files.

+++ libs/video/renderer/vid_render_vulkan.c
@@ -192,7 +192,7 @@ vulkan_R_Init (memhunk_t *hunk, struct plitem_s *config)
 		}
 		cache_data->size = head.size;
 		dstring_adjust (cache_data);
-		if (Qread (f, cache_data->str, head.size) != (int) head.size
+		if (Qread (f, cache_data->str, head.size) != head.size
 			|| Hash_Buffer (cache_data->str, cache_data->size) != head.hash) {
 			dstring_clear (cache_data);
 			goto cache_bail;
